ライブラリ更新

+++ core/src/commonMain/kotlin/io/github/mee1080/umasim/scenario/uaf/UafScenarioEvents.kt
@@ -105,7 +105,7 @@ class UafScenarioEvents : CommonScenarioEvents() {
                 it >= 5 -> 3
                 it >= 1 -> 1
                 else -> 0
-            }.toInt()
+            }
         }
 
         return addStatus(
